Cleaning up diagnostics
